requirements.txt
# Market Mayhem - Unified Requirements
# Python dependencies for the entire codebase
# Compatible with Python 3.8+ (tested with Python 3.11, 3.12, 3.13)

# FastAPI & Web Framework
fastapi>=0.104.0,<0.120.0
uvicorn[standard]>=0.24.0,<0.30.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
websockets>=11.0,<13.0
python-multipart>=0.0.6,<1.0.0

# LangChain & LangGraph
langgraph>=0.0.40,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-google-genai>=1.0.0,<3.0.0
langchain-community>=0.0.20,<1.0.0
langsmith>=0.0.80,<1.0.0

# Google AI
google-generativeai>=0.3.0,<1.0.0

# News & Data
tavily-python>=0.3.0,<1.0.0
yfinance>=0.2.0,<1.0.0

# Observability
opik>=0.1.0,<2.0.0

# Database
asyncpg>=0.29.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0

# Cache
redis>=5.0.0,<6.0.0
hiredis>=2.0.0,<3.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Utilities
httpx>=0.25.0,<1.0.0
python-json-logger>=2.0.0,<3.0.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
rich>=13.0.0,<14.0.0

# Development
black>=23.0.0,<25.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.0.0,<2.0.0
